fix: node-init should use docker if /etc/crictl.yaml not found
[ upstream commit 552c823f561149213807627cfbd724c39dbd8a10 ]

This script has several tests for what the container runtime situation
looks like to determine how best to restart the underlying containers
(going around the kubelet) so that the new networking configuration
can take effect.

The first test looks to see if the crictl config file is configured to use
docker, but if that file doesn't exist then it fails. I believe docker
is the default if this hasn't been configured at all so if that file
doesn't exist then use docker.

Fixes #12850

Signed-off-by: Nathan Bird <njbird@infiniteenergy.com>
Signed-off-by: John Fastabend <john.fastabend@gmail.com>